+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fix
 - Fix Proxy issue on PDP and for tracking requests
 - Handle ClientException for SSR (invalid credentials, API issue etc.)
+- Fix price slider filter closing
 
 ## [v5.2.2] - 2024.10.09
 ### Improve